% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_covs_base}
\alias{combine_covs_base}
\title{combine_covs_base}
\usage{
combine_covs_base(mn_covs1, mn_covs2, ns1, ns2)
}
\arguments{
\item{mn_covs1}{...}

\item{mn_covs2}{...}

\item{ns1}{...}

\item{ns2}{...}
}
\description{
combine_covs_base
}
