% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cov.R
\name{parallel_cov}
\alias{parallel_cov}
\title{Parallel covariance function}
\usage{
parallel_cov(x, splits = 2, cores = 2, cov_fun = "custom_multi_covar")
}
\arguments{
\item{x}{...}

\item{splits}{shows the number of splits of data that we want to combine  (should be of 2^n form)}

\item{cores}{number of cores that will be used for the parallel computation}

\item{cov_fun}{covariance function used on each split}
}
\description{
Parallel covariance function
}
