% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_quality.R
\name{variable_quality}
\alias{variable_quality}
\title{Measure variable quality based on replicate correlation}
\usage{
variable_quality(sample, variables, strata, replicates, key = NULL,
  batch = NULL)
}
\arguments{
\item{sample}{...}

\item{variables}{...}

\item{strata}{...}

\item{replicates}{...}

\item{key}{...}

\item{batch}{...}
}
\value{
data.frame of variable quality measurements
}
\description{
Measure variable quality based on replicate correlation
}

